move to new dir, typehints + docstrings, del old fns
